[Refactor] Change the GetStatistics query namespace
