+++ src/Sylius/Bundle/ApiBundle/QueryHandler/GetStatisticsHandler.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\QueryHandler\Admin;
+namespace Sylius\Bundle\ApiBundle\QueryHandler;
 
-use Sylius\Bundle\ApiBundle\Query\Admin\GetStatistics;
+use Sylius\Bundle\ApiBundle\Query\GetStatistics;
 use Sylius\Component\Channel\Context\ChannelNotFoundException;
 use Sylius\Component\Channel\Repository\ChannelRepositoryInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
