+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/query_handlers.xml
@@ -16,7 +16,7 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
 >
     <services>
-        <defaults public="true">
+        <defaults>
             <tag name="messenger.message_handler" bus="sylius.query_bus" />
         </defaults>
 
@@ -25,10 +25,9 @@
             <argument type="service" id="sylius.customer_statistics_provider" />
         </service>
 
-        <service id="Sylius\Bundle\ApiBundle\QueryHandler\Admin\GetStatisticsHandler">
+        <service id="Sylius\Bundle\ApiBundle\QueryHandler\GetStatisticsHandler">
             <argument type="service" id="Sylius\Component\Core\Statistics\Provider\StatisticsProviderInterface" />
             <argument type="service" id="sylius.repository.channel" />
-            <tag name="messenger.message_handler" bus="sylius.query_bus" />
         </service>
 
         <service id="Sylius\Bundle\ApiBundle\QueryHandler\GetAddressLogEntryCollectionHandler">
